[#175589933] refactoring
