+++ ts/features/bonus/bpd/components/paymentMethodActivationToggle/BancomatBpdToggle.tsx
@@ -1,8 +1,11 @@
 import * as React from "react";
 import pagoBancomatImage from "../../../../../../img/wallet/cards-icons/pagobancomat.png";
-import { PatchedWalletV2 } from "../../../../../types/pagopa";
+import {
+  EnableableFunctionsTypeEnum,
+  PatchedWalletV2
+} from "../../../../../types/pagopa";
 import { HPan } from "../../store/actions/paymentMethods";
-import { hasBpdCapability } from "../../utils";
+import { hasFunctionEnabled } from "../../../../../utils/walletv2";
 import PaymentMethodBpdToggle from "./base/PaymentMethodBpdToggle";
 
 type Props = {
@@ -18,7 +21,10 @@ export const BancomatBpdToggle: React.FunctionComponent<Props> = props => (
   <PaymentMethodBpdToggle
     hPan={props.card.info.hashPan as HPan}
     icon={pagoBancomatImage}
-    hasBpdCapability={hasBpdCapability(props.card)}
+    hasBpdCapability={hasFunctionEnabled(
+      props.card,
+      EnableableFunctionsTypeEnum.BPD
+    )}
     // TODO: load bank name and displayhere
     caption={`bank name here!`}
   />
