+++ ts/features/bonus/bpd/components/paymentMethodActivationToggle/BpdPaymentMethodToggleFactory.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { WalletTypeEnum } from "../../../../../../definitions/pagopa/bancomat/WalletV2";
+import { WalletTypeEnum } from "../../../../../../definitions/pagopa/walletv2/WalletV2";
 import { PatchedWalletV2 } from "../../../../../types/pagopa";
 import { BancomatBpdToggle } from "./BancomatBpdToggle";
 import { CardBpdToggle } from "./CardBpdToggle";
